requirements.txt
boto3>=1.26.0
mutagen>=1.46.0
google-genai>=1.0.0
pyyaml>=6.0
pytest>=7.0.0
